mat/qt02-mat-wannier-est/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qt02-mat-wannier-est"
version = "0.1.0"
description = "examples"
readme = "README.md"
requires-python = ">3.7,<3.9"
dependencies = [
    "jarvis-tools",
    "qiskit-aqua==0.7.5",
    "qiskit-terra==0.15.2",
    "qiskit-ignis==0.4.0",
    "numpy>=1.17,<1.20",
    "spglib<2.3",
    "toml"
]

[tool.setuptools]
packages = {find = {where = ["."]}}

[tool.black]
line-length = 88
[tool.ruff]
line-length = 88


